+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Updates Storybook and related npm packages to version `8.1.11`. Does not affect any DS component.
+- Adds `customLink` and `linkProps` to `Breadcrumbs` component.
 
 ## 3.1.7 (July 3, 2024)